Move combobox stories
